First  Commit
This is my initial commit to this project. I do a lot of testing on test2.py with ideas that I come up with and see if I can get the desired results. I will be continuing to use a lot of test files to see if I can retrieve the information that I want.